Customer Add New now has all the validations
Hey all, I just finished with the validations for Add New Customer. With
the State validation, I was checking to see if it was a US state, and if
so, did that Two letter state match with the two letter state that
corresponded with the Zip code that they entered. I think that is what
the requirement was. Jace, feel free to use the same code for checking
the state for Employee