Calculate Ticket Discounts on its way
I managed to get the base fare now and get one of the discounts (age
discount) to work. It's really messy right now, and I have to decide
where to make functions and where to make subs and public properties to
clean it up. I'm taking a break right now though and I plan to come back
to it in a couple of hours