Lots of fixes
The reservation page works properly now! (at least the bug we had
earlier is gone).
The entire reservation process works too (I think). I'm not staying up
another 45 minutes to check since its 2:10AM.
It works with one-way tickets but that's the easiest.
The reservation pages now just have to be hooked up to the Select Seats
(Which is made) and the stuff to calculate price/payments/mileage
paid/upgrades. After that, I think we're good to go. :)
Hopefully all of this will be done by tomorrow.